Fix menu button and keyboard input not working if ROM failed to load
