+++ test/ui/emulator/emulator_robot.dart
@@ -1,5 +1,4 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:nesd/ui/emulator/display.dart';
 import 'package:nesd/ui/emulator/emulator_widget.dart';
 
 import '../base_robot.dart';
@@ -12,6 +11,6 @@ class EmulatorRobot extends BaseRobot {
   }
 
   Future<void> tapMenu() async {
-    await goAsync(find.byKey(DisplayWidget.menuKey));
+    await goAsync(find.byKey(EmulatorWidget.menuKey));
   }
 }
